# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ICICLE AI Vector Service",
    "description": "Vector storage and retrieval service for the ICICLE AI tenant. Clients provide\ntheir own pre-computed embeddings; the service handles storage, filtered search\nand reranking.\n\n**Collections** are broad domains (e.g. `biology`, `chemistry`), each backed by\nits own Qdrant collection and HNSW index. **Topics** are optional sub-categories\nwithin a collection (e.g. `human`, `plant`).\n\n**User isolation.** Every request is authenticated with a Tapis access token via\nthe `X-Tapis-Token` header, and `user_id` is taken from the token's\n`tapis/username` claim — never from the request body. Collections are physically\nshared, but every read, write, search and delete is filtered by `user_id`, so\nusers never see or affect each other's data.\n",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/embeddings": {
      "post": {
        "tags": [
          "embeddings"
        ],
        "summary": "Store an embedding",
        "operationId": "store_embedding_v1_embeddings_post",
        "parameters": [
          {
            "name": "X-Tapis-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Tapis-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmbeddingCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmbeddingRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "TapisToken": []
          }
        ]
      }
    },
    "/v1/embeddings/bulk-delete": {
      "post": {
        "tags": [
          "embeddings"
        ],
        "summary": "Delete many embeddings in one collection",
        "description": "Delete by explicit ids, by topic/metadata predicate, or everything in the collection.\n\nUses POST rather than DELETE because the selector is a request body, which\nDELETE cannot carry portably. Only the caller's own points are ever removed.",
        "operationId": "bulk_delete_v1_embeddings_bulk_delete_post",
        "parameters": [
          {
            "name": "X-Tapis-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Tapis-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkDeleteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkDeleteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "TapisToken": []
          }
        ]
      }
    },
    "/v1/embeddings/{embedding_id}": {
      "get": {
        "tags": [
          "embeddings"
        ],
        "summary": "Get one embedding",
        "operationId": "read_embedding_v1_embeddings__embedding_id__get",
        "parameters": [
          {
            "name": "embedding_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Embedding Id"
            }
          },
          {
            "name": "collection",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Collection the embedding belongs to",
              "title": "Collection"
            },
            "description": "Collection the embedding belongs to"
          },
          {
            "name": "X-Tapis-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Tapis-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmbeddingRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "TapisToken": []
          }
        ]
      },
      "put": {
        "tags": [
          "embeddings"
        ],
        "summary": "Update an embedding",
        "operationId": "update_user_embedding_v1_embeddings__embedding_id__put",
        "parameters": [
          {
            "name": "embedding_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Embedding Id"
            }
          },
          {
            "name": "collection",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Collection the embedding belongs to",
              "title": "Collection"
            },
            "description": "Collection the embedding belongs to"
          },
          {
            "name": "X-Tapis-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Tapis-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmbeddingUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmbeddingRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "TapisToken": []
          }
        ]
      },
      "delete": {
        "tags": [
          "embeddings"
        ],
        "summary": "Delete an embedding",
        "operationId": "delete_user_embedding_v1_embeddings__embedding_id__delete",
        "parameters": [
          {
            "name": "embedding_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Embedding Id"
            }
          },
          {
            "name": "collection",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Collection the embedding belongs to",
              "title": "Collection"
            },
            "description": "Collection the embedding belongs to"
          },
          {
            "name": "X-Tapis-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Tapis-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "TapisToken": []
          }
        ]
      }
    },
...
    "/v1/collections/{collection}/embeddings": {
      "get": {
        "tags": [
          "collections"
        ],
        "summary": "List your embeddings in a collection",
        "description": "Cursor-paginated listing. Returns metadata only — never vectors.",
        "operationId": "list_embeddings_v1_collections__collection__embeddings_get",
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Collection"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "description": "Page size",
              "default": 50,
              "title": "Limit"
            },
            "description": "Page size"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Cursor from a previous page's next_offset",
              "title": "Offset"
            },
            "description": "Cursor from a previous page's next_offset"
          },
          {
            "name": "topic",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only embeddings with this topic",
              "title": "Topic"
            },
            "description": "Only embeddings with this topic"
          },
          {
            "name": "X-Tapis-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Tapis-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmbeddingListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "TapisToken": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "BulkDeleteRequest": {
        "properties": {
          "collection": {
            "type": "string",
            "title": "Collection"
          },
          "ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ids"
          },
          "topic": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic"
          },
          "filter": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MetadataFilter"
              },
              {
                "type": "null"
              }
            ]
          },
          "all": {
            "type": "boolean",
            "title": "All",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "collection"
        ],
        "title": "BulkDeleteRequest",
        "description": "Delete many of the caller's embeddings in one collection.\n\nExactly one selector must be given: explicit ``ids``, or a\n``topic``/``filter`` predicate. ``all=true`` deletes every point the caller\nowns in the collection."
      },
      "BulkDeleteResponse": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "collection": {
            "type": "string",
            "title": "Collection"
          },
          "deleted": {
            "type": "integer",
            "title": "Deleted"
          },
          "collection_dropped": {
            "type": "boolean",
            "title": "Collection Dropped",
            "description": "True when the underlying Qdrant collection was removed because no points from any user remained. Always false unless the deployment sets DROP_EMPTY_COLLECTIONS=true.",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "collection",
          "deleted"
        ],
        "title": "BulkDeleteResponse"
      },
...
      "DeleteResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "deleted": {
            "type": "boolean",
            "title": "Deleted"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "deleted"
        ],
        "title": "DeleteResponse"
      },
      "EmbeddingCreate": {
        "properties": {
          "embedding": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Embedding"
          },
          "collection": {
            "type": "string",
            "title": "Collection"
          },
          "topic": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic"
          },
          "chunks": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Chunks"
          },
          "token_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Token Ids"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          },
          "embedding_model": {
            "type": "string",
            "title": "Embedding Model"
          }
        },
        "type": "object",
        "required": [
          "embedding",
          "collection",
          "chunks",
          "embedding_model"
        ],
        "title": "EmbeddingCreate"
      },
      "EmbeddingListResponse": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "collection": {
            "type": "string",
            "title": "Collection"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "next_offset": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Offset",
            "description": "Pass as ?offset= to fetch the next page. Null on the last page."
          },
          "embeddings": {
            "items": {
              "$ref": "#/components/schemas/EmbeddingRecord"
            },
            "type": "array",
            "title": "Embeddings"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "collection",
          "count",
          "embeddings"
        ],
        "title": "EmbeddingListResponse",
        "description": "A page of the caller's embeddings in one collection (metadata only)."
      },
      "EmbeddingRecord": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "collection": {
            "type": "string",
            "title": "Collection"
          },
          "topic": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic"
          },
          "vector_dim": {
            "type": "integer",
            "title": "Vector Dim"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "embedding_model": {
            "type": "string",
            "title": "Embedding Model"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "collection",
          "vector_dim",
          "created_at",
          "updated_at",
          "embedding_model"
        ],
        "title": "EmbeddingRecord"
      },
      "EmbeddingUpdate": {
        "properties": {
          "embedding": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Embedding"
          },
          "topic": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic"
          },
          "chunks": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chunks"
          },
          "token_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token Ids"
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata"
          },
          "embedding_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Embedding Model"
          }
        },
        "type": "object",
        "title": "EmbeddingUpdate"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MetadataFilter": {
        "properties": {
          "conditions": {
            "additionalProperties": true,
            "type": "object",
            "title": "Conditions",
            "description": "Key-value pairs to match against the metadata payload field"
          }
        },
        "type": "object",
        "title": "MetadataFilter",
        "description": "Filter results by metadata fields.\n\nEach key-value pair matches against the nested ``metadata`` payload field.\nAll conditions are ANDed together. Values can be:\n  - a string/number/bool for exact match\n  - a list for \"any of\" match\n\nKeys are namespaced under ``metadata.`` when translated to Qdrant, so this\ncannot be used to address a top-level payload field such as ``user_id``."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "TapisToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Tapis-Token",
        "description": "ICICLE AI tenant Tapis access token. The username is read from the tapis/username claim and used to isolate your data."
      }
    }
  },
  "servers": [
    {
      "url": "https://icicleaivecserver.pods.icicleai.tapis.io",
      "description": "Production (ICICLE AI Tapis pod)"
    },
    {
      "url": "http://localhost:8000",
      "description": "Local development"
    }
  ],
...
}